training/requirements.txt
# Model Training
torch==2.1.0
transformers==4.35.0
accelerate==0.25.0
datasets==2.14.7
tokenizers==0.15.0
sentencepiece==0.1.99

# Training Optimization
deepspeed==0.12.3
bitsandbytes==0.41.3
flash-attn==2.4.2

# Data Processing
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1
pyarrow==14.0.1

# Monitoring
wandb==0.16.0
tensorboard==2.15.1

# Utilities
pyyaml==6.0.1
tqdm==4.66.1

